survey.py api 분해 (#118)
